+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
 
 - Rework handling of running from a source checkout, fixes issues with example files
   when `urlwatch` was run as `/usr/sbin/urlwatch`, e.g. on Void Linux (fixes #755)
+- Add support for docutils >= 0.18, which deprecated `frontend.OptionParser` (fixes #754)
 
 ## [2.26] -- 2023-04-11